refactor(web): extract FaceThumbGrid and PurgeFacesDialog into shared components
