+++ apps/web/src/components/people/PurgeFacesDialog.tsx
@@ -1,0 +1,77 @@
+import { useState } from 'react';
+import {
+  Alert,
+  Button,
+  CircularProgress,
+  Dialog,
+  DialogActions,
+  DialogContent,
+  DialogTitle,
+  Typography,
+} from '@mui/material';
+
+interface PurgeFacesDialogProps {
+  open: boolean;
+  count: number;
+  onClose: () => void;
+  onConfirm: () => Promise<void>;
+}
+
+export function PurgeFacesDialog({ open, count, onClose, onConfirm }: PurgeFacesDialogProps) {
+  const [deleting, setDeleting] = useState(false);
+  const [error, setError] = useState<string | null>(null);
+
+  const handleConfirm = async () => {
+    setDeleting(true);
+    setError(null);
+    try {
+      await onConfirm();
+      onClose();
+    } catch (err) {
+      setError(err instanceof Error ? err.message : 'Deletion failed. Please try again.');
+    } finally {
+      setDeleting(false);
+    }
+  };
+
+  const handleClose = () => {
+    if (deleting) return;
+    setError(null);
+    onClose();
+  };
+
+  return (
+    <Dialog open={open} onClose={handleClose} maxWidth="xs" fullWidth>
+      <DialogTitle>Delete permanently?</DialogTitle>
+      <DialogContent>
+        <Typography variant="body2" sx={{ mb: 1 }}>
+          This permanently removes{' '}
+          <strong>
+            {count} face{count !== 1 ? 's' : ''}
+          </strong>{' '}
+          and their biometric data. <strong>Your photos are NOT deleted.</strong> This cannot be
+          undone.
+        </Typography>
+        {error && (
+          <Alert severity="error" sx={{ mt: 2 }}>
+            {error}
+          </Alert>
+        )}
+      </DialogContent>
+      <DialogActions>
+        <Button onClick={handleClose} disabled={deleting}>
+          Cancel
+        </Button>
+        <Button
+          variant="contained"
+          color="error"
+          onClick={() => void handleConfirm()}
+          disabled={deleting}
+          startIcon={deleting ? <CircularProgress size={16} /> : undefined}
+        >
+          {deleting ? 'Deleting…' : 'Delete permanently'}
+        </Button>
+      </DialogActions>
+    </Dialog>
+  );
+}
